Renamed MultiSelect to MultiSelectField.
