+++ frontend/admin/src/js/components/pool/CreatePool.tsx
@@ -4,7 +4,7 @@ import { useIntl } from "react-intl";
 import { toast } from "react-toastify";
 import {
   Input,
-  MultiSelect,
+  MultiSelectField,
   Select,
   Submit,
   TextArea,
@@ -232,7 +232,7 @@ export const CreatePoolForm: React.FunctionComponent<CreatePoolFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="classifications"
               label={intl.formatMessage({
                 defaultMessage: "Classifications",
@@ -250,7 +250,7 @@ export const CreatePoolForm: React.FunctionComponent<CreatePoolFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="assetCriteria"
               label={intl.formatMessage({
                 defaultMessage: "Asset Criteria",
@@ -268,7 +268,7 @@ export const CreatePoolForm: React.FunctionComponent<CreatePoolFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="essentialCriteria"
               label={intl.formatMessage({
                 defaultMessage: "Essential Criteria",
@@ -286,7 +286,7 @@ export const CreatePoolForm: React.FunctionComponent<CreatePoolFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="operationalRequirements"
               name="operationalRequirements"
               label={intl.formatMessage({
