+++ frontend/admin/src/js/components/skill/CreateSkill.tsx
@@ -3,7 +3,7 @@ import { useIntl } from "react-intl";
 import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import sortBy from "lodash/sortBy";
 import { toast } from "react-toastify";
-import { Input, TextArea, Submit, MultiSelect } from "@common/components/form";
+import { Input, TextArea, Submit, MultiSelectField } from "@common/components/form";
 import { getLocale } from "@common/helpers/localize";
 import { notEmpty } from "@common/helpers/util";
 import { navigate } from "@common/helpers/router";
@@ -231,7 +231,7 @@ export const CreateSkillForm: React.FunctionComponent<CreateSkillFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="families"
               name="families"
               label={intl.formatMessage({
