+++ frontend/admin/src/js/components/skill/UpdateSkill.tsx
@@ -4,7 +4,7 @@ import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import pick from "lodash/pick";
 import sortBy from "lodash/sortBy";
 import { toast } from "react-toastify";
-import { Submit, Input, MultiSelect, TextArea } from "@common/components/form";
+import { Submit, Input, MultiSelectField, TextArea } from "@common/components/form";
 import { notEmpty } from "@common/helpers/util";
 import { navigate } from "@common/helpers/router";
 import { getLocale } from "@common/helpers/localize";
@@ -222,7 +222,7 @@ export const UpdateSkillForm: React.FunctionComponent<UpdateSkillFormProps> = ({
                 required: intl.formatMessage(errorMessages.required),
               }}
             />
-            <MultiSelect
+            <MultiSelectField
               id="families"
               name="families"
               label={intl.formatMessage({
