+++ frontend/common/src/components/form/MultiSelectField/MultiSelectFieldV2.stories.tsx
