+++ frontend/common/src/components/form/index.ts
@@ -2,7 +2,7 @@ export { BasicForm } from "./BasicForm";
 export { default as Checkbox } from "./Checkbox";
 export { default as Checklist } from "./Checklist";
 export { default as Input } from "./Input";
-export { default as MultiSelect } from "./MultiSelect";
+export { default as MultiSelectField } from "./MultiSelectField";
 export { default as Radio } from "./Radio";
 export { default as RadioGroup } from "./RadioGroup";
 export { default as Select } from "./Select";
